refactor: based on comments
