+++ app/src/test/kotlin/com/waz/zclient/auth/registration/activation/ActivationRemoteDataSourceTest.kt
@@ -4,7 +4,6 @@ import com.waz.zclient.UnitTest
 import com.waz.zclient.capture
 import com.waz.zclient.core.network.NetworkHandler
 import kotlinx.coroutines.*
-import kotlinx.coroutines.test.runBlockingTest
 import org.amshove.kluent.shouldBe
 import org.junit.Before
 import org.junit.Test
@@ -27,13 +26,11 @@ class ActivationRemoteDataSourceTest : UnitTest() {
     @Mock
     private lateinit var networkHandler: NetworkHandler
 
-
     @Mock
     private lateinit var emptyResponse: Response<Unit>
 
     @Captor
     private lateinit var sendEmailActivationCodeRequestCaptor: ArgumentCaptor<SendActivationCodeRequest>
-
 
     @Before
     fun setUp() {
@@ -71,7 +68,6 @@ class ActivationRemoteDataSourceTest : UnitTest() {
 
         activationRemoteDataSource.sendEmailActivationCode(TEST_EMAIL).isLeft shouldBe true
     }
-
 
     @Test(expected = CancellationException::class)
     fun `Given  sendActivationCode()() is called, when api response is cancelled, then return an error`() = runBlocking {
